Adds a call where your mother blended a sandwich to the phone traps (#92)
* Adds a call where your mother blended a sandwich to the phone traps

Credits to Palex for the idea.

Reference description: There was a hololive streamer name Ouro Kronii that blended subway sandwiches and drank them for dinner before. I umm... don't think I need to go into detail on that. However, this call is an adaptation of that, that doesn't go into detail, and is instead your mom calling about having done so.

Reference:
(WARNING: Looks nasty.)
https://www.youtube.com/watch?v=ICN7KlZ-BTs

* Added trailing comma

* Fix command

---------

Co-authored-by: James White <james_white_d@hotmail.com>