Rework level scaling sphere calculation (#370)
This simplifies the level scaling sphere calculation, reducing time
spent in output generation from 17% to 2.6% when using the fuzzer
profiler hook (slightly unscientific but it's a good estimation and
is a stable percentage across runs).

This removes 2 levels of loops which makes it a lot easier to reason
about, we only calculate distances once per sphere instead of once per
event, and it's overall just more efficient.

The region check might seem useless since location reachability also
checks for region access which is cached, but shear overhead of
calling methods instead of skipping locations makes it twice as slow...